Fix Rally context external networks
This patch fixes the issue with the gateway and CIDR
ranges in rally context external networks that
are used in trunk dynamic workloads and VM dynamic
workloads. Previously, rally was creating subnets
with different CIDRs than the network interfaces on the
undercloud. Also, we were providing an incorrect IP
as the gateway for the network interface. These two issues
are fixed in this patch.

Change-Id: I5d14d2ed2ad4b54fc82b37009223d16fee5d9454