stressng : Improve error handling
The browbeat stressng workload reports an error in the logs
while running the stressng command, even though the output of
the stressng command reports that it has executed successfully[1].
This happens in all iterations. Also, the iteration ends with END: OK
even if there happens to be a genuine error in executing the command.

This patch fixes these 2 issues by checking the exitcode of the stressng
command and raising an exception if it is nonzero.

[1] https://privatebin.corp.redhat.com/?dc09ee175a9ee377#37vYuNmR4CF2Hc3ghjcggXeW1EHY5WXDH5PB9Wb5sBoZ

Change-Id: I0d49ee6f80e4ae9b2946c67e8689daaccb200b1d